+++ src/types.py
@@ -9,8 +9,7 @@ from pydantic import BaseModel
 
 class BigQueryClientConfig(BaseModel):
     """
-    The class implements a Pydantic type for a BigQuery Client
-    configuration.
+    BigQuery Client configuration
 
     Attributes:
         project_id (String): The Google Cloud project ID, which is
@@ -22,8 +21,7 @@ class BigQueryClientConfig(BaseModel):
 
 class BigQueryQueryParameter(BaseModel):
     """
-    The class implements a Pydantic type for a BigQuery Query
-    parameter
+    BigQuery Query parameter
 
     Attributes:
         name (String): Parameter name
@@ -39,8 +37,7 @@ class BigQueryQueryParameter(BaseModel):
 
 class BigQueryQueryConfig(BaseModel):
     """
-    The class implements a Pydantic type for a BigQuery Query
-    configuration
+    BigQuery Query configuration
 
     Attributes:
         query_path (String): Query file path
@@ -62,3 +59,67 @@ class BigQueryQueryConfig(BaseModel):
             (Integer): Number of non-None attributes
         """
         return sum(1 for field, value in self.__dict__.items() if value is not None)
+
+
+class SentenceTransformersConfig(BaseModel):
+    """
+    Configuration for embedding generation with SentenceTransformers library
+
+    Attributes:
+        model_name (String): The name of the model to use
+        numpy_tensor (Boolean): Output tensor to be a numpy array
+    """
+
+    model_name: str
+    numpy_tensor: bool = False
+
+
+class EmbeddingsConfig(BaseModel):
+    """
+    Configuration for embedding generation model
+
+    Attributes:
+        method (String): The embedding approach to use (e.g., SentenceTransformer)
+        embedding_model_config (Union[SentenceTransformersConfig]): Model configuration
+    """
+
+    method: str
+    embedding_model_config: Union[SentenceTransformersConfig]
+
+
+class PCAConfig(BaseModel):
+    """
+    Configuration for a PCA model
+
+    Attributes:
+        n_components (Integer): Number of components
+    """
+
+    n_components: int
+
+
+class CompressEmbeddingsConfig(BaseModel):
+    """
+    Configuration for compressing embeddings model
+
+    Attributes:
+        method (String): The compress approach to use (e.g., PCA)
+        compress_model_config (Union[PCAConfig]): Model configuration
+    """
+
+    method: str
+    compress_model_config: Union[PCAConfig]
+
+
+class EncodingTextConfig(BaseModel):
+    """
+    Configuration to encode Text and compress them into a lower
+    dimensional vector
+
+    Attributes:
+        embeddings_config (EmbeddingsConfig): Configuration for embedding generation
+        compress_embeddings_config (CompressEmbeddingsConfig): Configuration for embedding compression
+    """
+
+    embeddings_config: EmbeddingsConfig
+    compress_embeddings_config: CompressEmbeddingsConfig